New component: games
* src/components/games/CMakeLists.txt:
* src/components/games/games.c:
* src/components/games/games.h:

  New files.

Pull in the provided skeleton for game protocol communication.  This
is going to require adjustment to the fact that we're using the base-2
encoding as our working representation of the board.

Consider:
Maybe rename tictactoe_board_t to tictacoe_t (the name used for the
sparse representation in the skeleton).  Assuming everything in the
skeleton uses the tictac_get, tictac_set, etc. API and does not make
assumptions based on the in memory size of the board, such a change
should make our more memory efficient board representation a drop in
replacement.  If this assumption holds, this renaming will only
require changes in our code/code already adjusted in the tictactoe
component, and should be achievable with a simple search-replace.