Change tictactoe_board_t to tictactoe_t
* src/components/tictactoe/tictactoe.h (tictactoe_board_t): Renamed to
  tictactoe_t.

  Users adjusted.

As noted in 1eb220dade8a02cc5de4283c4ad33f9a9cc2d0b6 the simplest way
to integrate into the rest of the skeleton was to simply replace the
implementation of the sparse board representation.  Since we provide
the same interface to the base-2 representation no code adjustments
were required in the game component.