Remove the VT100 codes
* src/components/games/games.c (tictac_display_board): Remove the
  VT100 codes.

The VT100 codes kept the game board at the top of the screen, this was
intended to overwrite the board in place instead of constantly writing
a new one at the bottom.  The thought is good, but with all the other
output being written to the serial port it is not the most readable.